\name{data1}
\alias{data1}
\docType{data}
\title{
data1: Sampaio (2010): page 134
}
\description{
Quantitative treatments in completely randomized design.
}
\usage{data(data1)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{treatment}}{a numeric vector}
    \item{\code{gain}}{a numeric vector}
  }
}

\references{
SAMPAIO, I. B. M. Estatistica aplicada a experimentacao animal. 3nd Edition. Belo Horizonte: Editora FEPMVZ, Fundacao de Ensino e Pesquisa em Medicina Veterinaria e Zootecnia, 2010. 264p.
}
\examples{
data(data1)
summary(data1)
}

