\name{data5}
\alias{data5}
\docType{data}
\title{
data5: fictional example
}
\description{
Quantitative treatments and three response variable.
}
\usage{data(data5)}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{treatments}}{a numeric vector}
    \item{\code{variable1}}{a numeric vector}
    \item{\code{variable2}}{a numeric vector}
    \item{\code{variable3}}{a numeric vector}
  }
}

\examples{
data(data5)
summary(data5)
}

