% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epr_df.R
\docType{data}
\name{epr_df}
\alias{epr_df}
\alias{epr_status_codes_df}
\title{EPR - Ethnic Power Relations}
\format{
\itemize{
\item Variables:
\itemize{
\item \code{gwid}
\itemize{
\item \href{http://ksgleditsch.com/data-4.html}{Gleditsch and Ward State ID}
}
\item \code{statename}
\item \code{from}
\item \code{to}
\item \code{group}
\item \code{groupid}
\item \code{gwgroupid}
\item \code{umbrella}
\item \code{size}
\item \code{status}
\itemize{
\item The group rules alone:
\itemize{
\item \code{"MONOPOLY"}
\item Elite members hold monopoly power in the executive to the exclusion
of members of all other ethnic groups.
\item \code{"DOMINANCE"}
\itemize{
\item Elite members of the group hold dominant power in the executive but
there is some limited inclusion of “token” members of other groups
who however do not have real influence on decision making.
}
}
\item The group shares power:
\itemize{
\item \code{"SENIOR PARTNER"}
\itemize{
\item Representatives of the group participate as senior partners in a
formal or informal power-sharing arrangement. By power sharing, we mean
any arrangement that divides executive power among leaders who claim
to represent particular ethnic groups and who have real influence on
political decision making.
}
\item \code{"JUNIOR PARTNER"}
\itemize{
\item Representatives participate as junior partners in government.
\itemize{
\item The choice between senior and junior partner depends on the group’s
absolute influence in the executive - that is, irrespective of group
size -, measured by the number and importance of the positions
controlled by group members.
}
}
}
\item The group is excluded:
\itemize{
\item \code{"POWERLESS"}
\itemize{
\item Elite representatives hold no political power (or do not have
influence on decision making) at the national level of executive
power - although without being explicitly discriminated against.
}
\item \code{"DISCRIMINATION"}
\itemize{
\item Group members are subjected to active, intentional, and targeted
discrimination by the state, with the intent of excluding them from
political power. Such active discrimination can be either formal or
informal, but always refers to the domain of public politics
(excluding discrimination in the socio-economic sphere).
}
\item \code{"SELF-EXCLUSION"}
\itemize{
\item The special category of self-exclusion applies to groups that have
excluded themselves from central state power, in the sense that
they control a particular territory of the state which they have
declared independent from the central government.
\itemize{
\item This category was labeled “separatist autonomy” in previous EPR versions.
}
}
}
}
\item \code{reg_aut}
}
}


}
\source{
\itemize{
\item \href{https://icr.ethz.ch/data/epr/core/}{EPR - Ethnic Power Relations}
\itemize{
\item \href{https://icr.ethz.ch/data/epr/core/EPR-2019.csv}{Data (CSV)}
\item \href{https://icr.ethz.ch/data/epr/core/EPR_2019_Codebook_EPR.pdf}{Codebook}
}
}
}
\usage{
epr_df

epr_status_codes_df
}
\description{
\emph{The Ethnic Power Relations (EPR) Core Dataset 2019 is based on the original
dataset \href{https://icr.ethz.ch/data/epr/core/v1.1/}{EPR v.1.1} and identifies
all politically relevant ethnic groups and their access to state power in
every country of the world from 1946 to 2017. It includes annual data on
over 800 groups and codes the degree to which their representatives held
executive-level state power—from total control of the government to overt
political discrimination.}
}
\examples{
epr_df

epr_status_codes_df 

}
\references{
\itemize{
\item Vogt, Manuel, Nils-Christian Bormann, Seraina Rüegger, Lars-Erik Cederman, Philipp Hunziker, and Luc Girardin. 2015. “\href{https://icr.ethz.ch/publications/integrating-data-on-ethnicity-geography-and-conflict/}{Integrating Data on Ethnicity, Geography, and Conflict: The Ethnic Power Relations Data Set Family}.” Journal of Conflict Resolution 59(7): 1327–42.
\item Cederman, Lars-Erik, Andreas Wimmer, and Brian Min. 2010. “\href{https://icr.ethz.ch/publications/why-do-ethnic-groups-rebel/}{Why Do Ethnic Groups Rebel? New Data and Analysis}.” World Politics 62(1): 87–119.
}
}
\keyword{datasets}
